Reorganize distributed tree files
