+++ test/tstDetailsDistributedTreeImpl.cpp
@@ -11,12 +11,12 @@
 
 #include "ArborX_EnableDeviceTypes.hpp" // ARBORX_DEVICE_TYPES
 #include "ArborX_EnableViewComparison.hpp"
-#include <ArborX_DetailsDistributedTreeImpl.hpp>
+#include <ArborX_DetailsDistributedTreeUtils.hpp>
 #include <ArborX_DetailsKokkosExtStdAlgorithms.hpp>
 
 #include <boost/test/unit_test.hpp>
 
-#define BOOST_TEST_MODULE DetailsDistributedTreeImpl
+#define BOOST_TEST_MODULE DetailsDistributedTree
 
 #include <set>
 
@@ -68,8 +68,8 @@ BOOST_AUTO_TEST_CASE_TEMPLATE(sort_results, DeviceType, ARBORX_DEVICE_TYPES)
   Kokkos::deep_copy(ranks, ranks_host);
 
   using ExecutionSpace = typename DeviceType::execution_space;
-  ArborX::Details::DistributedTreeImpl<DeviceType>::sortResults(
-      ExecutionSpace{}, ids, results, ranks);
+  ArborX::Details::DistributedTree::sortResults(ExecutionSpace{}, ids, results,
+                                                ranks);
 
   // COMMENT: ids are untouched
   Kokkos::deep_copy(ids_host, ids);
@@ -85,10 +85,9 @@ BOOST_AUTO_TEST_CASE_TEMPLATE(sort_results, DeviceType, ARBORX_DEVICE_TYPES)
     }
 
   Kokkos::View<int *, DeviceType> not_sized_properly("", m);
-  BOOST_CHECK_THROW(
-      ArborX::Details::DistributedTreeImpl<DeviceType>::sortResults(
-          ExecutionSpace{}, ids, not_sized_properly),
-      ArborX::SearchException);
+  BOOST_CHECK_THROW(ArborX::Details::DistributedTree::sortResults(
+                        ExecutionSpace{}, ids, not_sized_properly),
+                    ArborX::SearchException);
 }
 
 BOOST_AUTO_TEST_CASE_TEMPLATE(sort_results_2d, DeviceType, ARBORX_DEVICE_TYPES)
@@ -110,8 +109,8 @@ BOOST_AUTO_TEST_CASE_TEMPLATE(sort_results_2d, DeviceType, ARBORX_DEVICE_TYPES)
   Kokkos::deep_copy(results_2d, results_2d_host);
 
   using ExecutionSpace = typename DeviceType::execution_space;
-  ArborX::Details::DistributedTreeImpl<DeviceType>::sortResults(
-      ExecutionSpace{}, ids, results_2d);
+  ArborX::Details::DistributedTree::sortResults(ExecutionSpace{}, ids,
+                                                results_2d);
 
   Kokkos::deep_copy(results_2d_host, results_2d);
   for (int i = 0; i < 5; ++i)
@@ -147,8 +146,8 @@ BOOST_AUTO_TEST_CASE_TEMPLATE(sort_results_3d, DeviceType, ARBORX_DEVICE_TYPES)
   Kokkos::deep_copy(results_3d, results_3d_host);
 
   using ExecutionSpace = typename DeviceType::execution_space;
-  ArborX::Details::DistributedTreeImpl<DeviceType>::sortResults(
-      ExecutionSpace{}, ids, results_3d);
+  ArborX::Details::DistributedTree::sortResults(ExecutionSpace{}, ids,
+                                                results_3d);
 
   Kokkos::deep_copy(results_3d_host, results_3d);
   for (int i = 0; i < 5; ++i)
@@ -177,8 +176,8 @@ BOOST_AUTO_TEST_CASE_TEMPLATE(count_results, DeviceType, ARBORX_DEVICE_TYPES)
   Kokkos::View<int *, DeviceType> offset("offset", m);
 
   using ExecutionSpace = typename DeviceType::execution_space;
-  ArborX::Details::DistributedTreeImpl<DeviceType>::countResults(
-      ExecutionSpace{}, m, ids, offset);
+  ArborX::Details::DistributedTree::countResults(ExecutionSpace{}, m, ids,
+                                                 offset);
 
   auto offset_host = Kokkos::create_mirror_view(offset);
   Kokkos::deep_copy(offset_host, offset);
@@ -359,7 +358,7 @@ struct Helper
     // NOTE here we assume that the reference solution is sized properly
     auto v_imp = Kokkos::create_mirror(typename View3::memory_space(), v_ref);
 
-    ArborX::Details::DistributedTreeImpl<DeviceType>::sendAcrossNetwork(
+    ArborX::Details::DistributedTree::sendAcrossNetwork(
         typename DeviceType::execution_space{}, distributor, v_exp, v_imp);
 
     auto v_imp_host = Kokkos::create_mirror_view(v_imp);
